A first attempt at changing a bunch of instrumented tests into regular junit.
