+++ kin-sdk/kin-sdk-lib/build.gradle
@@ -55,6 +55,10 @@ dependencies {
     testImplementation 'com.squareup.okhttp3:mockwebserver:3.9.1'
     testImplementation 'org.robolectric:robolectric:4.3'
     testImplementation 'com.google.code.gson:gson:2.8.2'
+    testImplementation 'org.json:json:20180813'
+    testImplementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
+    testImplementation project(':kin-sdk:kin-base')
 
     androidTestImplementation 'junit:junit:4.12'
     androidTestImplementation 'org.mockito:mockito-core:2.10.0'
