+++ kin-sdk/kin-sdk-lib/src/androidTest/java/kin/sdk/FakeKeyStore.java
