+++ kin-sdk/kin-sdk-lib/src/main/java/kin/sdk/KeyStoreImpl.java
@@ -2,15 +2,17 @@ package kin.sdk;
 
 
 import android.support.annotation.NonNull;
-import android.text.TextUtils;
+import java.util.ArrayList;
+import java.util.List;
 import kin.base.KeyPair;
-import kin.sdk.exception.*;
+import kin.sdk.exception.CorruptedDataException;
+import kin.sdk.exception.CreateAccountException;
+import kin.sdk.exception.CryptoException;
+import kin.sdk.exception.DeleteAccountException;
+import kin.sdk.exception.LoadAccountException;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-
-import java.util.ArrayList;
-import java.util.List;
 
 class KeyStoreImpl implements KeyStore {
 
@@ -90,12 +92,12 @@ class KeyStoreImpl implements KeyStore {
         return addKeyPairToStorage(KeyPair.random());
     }
 
-    private KeyPair addKeyPairToStorage(KeyPair newKeyPair) throws CreateAccountException {
+    protected KeyPair addKeyPairToStorage(KeyPair newKeyPair) throws CreateAccountException {
         try {
             String encryptedSeed = String.valueOf(newKeyPair.getSecretSeed());
             String publicKey = newKeyPair.getAccountId();
             String accounts = store.getString(STORE_KEY_ACCOUNTS);
-            if (TextUtils.isEmpty(accounts) || !accounts.contains(publicKey)) {
+            if (Utils.isEmpty(accounts) || !accounts.contains(publicKey)) {
                 JSONObject accountsJson = addKeyPairToAccountsJson(encryptedSeed, publicKey);
                 store.saveString(STORE_KEY_ACCOUNTS, accountsJson.toString());
             }
