+++ kin-sdk/kin-sdk-lib/src/main/java/kin/sdk/KinClient.java
@@ -25,17 +25,10 @@ import static kin.sdk.Utils.checkNotNull;
 public class KinClient {
 
     private static final String STORE_NAME_PREFIX = "KinKeyStore_";
-    private final Environment environment;
-    private final KeyStore keyStore;
-    private final TransactionSender transactionSender;
-    private final AccountInfoRetriever accountInfoRetriever;
-    private final GeneralBlockchainInfoRetrieverImpl generalBlockchainInfoRetriever;
-    private final BlockchainEventsCreator blockchainEventsCreator;
-    private final BackupRestore backupRestore;
-    private final String appId;
+
+    private final KinClientInternal kinClientInternal;
     private final String storeKey;
-    @NonNull
-    private List<KinAccountImpl> kinAccounts = new ArrayList<>(1);
+    private final BackupRestore backupRestore;
 
     /**
      * For more details please look at {@link #KinClient(Context context,Environment environment, String appId, String storeKey)}
@@ -55,87 +48,25 @@ public class KinClient {
      */
     public KinClient(@NonNull Context context, @NonNull Environment environment, @NonNull String appId, @NonNull String storeKey) {
         checkNotNull(storeKey, "storeKey");
-        checkNotNull(context, "context");
-        checkNotNull(environment, "environment");
-        validateAppId(appId);
-        this.environment = environment;
-        this.backupRestore = new BackupRestoreImpl();
-        Server server = initServer();
-        this.appId = appId;
         this.storeKey = storeKey;
-        keyStore = initKeyStore(context.getApplicationContext(), storeKey);
-        transactionSender = new TransactionSender(server, appId);
-        accountInfoRetriever = new AccountInfoRetriever(server);
-        generalBlockchainInfoRetriever = new GeneralBlockchainInfoRetrieverImpl(server);
-        blockchainEventsCreator = new BlockchainEventsCreator(server);
-        loadAccounts();
+        backupRestore = new BackupRestoreImpl();
+        kinClientInternal = new KinClientInternal(createKeyStore(context, storeKey), environment, appId, backupRestore);
     }
 
     @VisibleForTesting
     KinClient(Environment environment, KeyStore keyStore, TransactionSender transactionSender,
         AccountInfoRetriever accountInfoRetriever, GeneralBlockchainInfoRetrieverImpl generalBlockchainInfoRetriever,
         BlockchainEventsCreator blockchainEventsCreator, BackupRestore backupRestore, String appId, String storeKey) {
-        this.environment = environment;
-        this.keyStore = keyStore;
-        this.transactionSender = transactionSender;
-        this.accountInfoRetriever = accountInfoRetriever;
-        this.generalBlockchainInfoRetriever = generalBlockchainInfoRetriever;
-        this.blockchainEventsCreator = blockchainEventsCreator;
-        this.backupRestore = backupRestore;
-        this.appId = appId;
         this.storeKey = storeKey;
-        loadAccounts();
+        this.backupRestore = backupRestore;
+        kinClientInternal = new KinClientInternal(environment, keyStore, transactionSender, accountInfoRetriever,
+            generalBlockchainInfoRetriever, blockchainEventsCreator, backupRestore, appId);
     }
 
-    private Server initServer() {
-        Network.use(environment.getNetwork());
-        return new Server(environment.getNetworkUrl(), new KinOkHttpClientFactory(BuildConfig.VERSION_NAME).client);
-    }
-
-    private KeyStore initKeyStore(Context context, String id) {
+    private KeyStore createKeyStore(Context context, String id) {
         SharedPrefStore store = new SharedPrefStore(
             context.getSharedPreferences(STORE_NAME_PREFIX + id, Context.MODE_PRIVATE));
         return new KeyStoreImpl(store, backupRestore);
-    }
-
-    private void loadAccounts() {
-        List<KeyPair> accounts = null;
-        try {
-            accounts = keyStore.loadAccounts();
-        } catch (LoadAccountException e) {
-            e.printStackTrace();
-        }
-        if (accounts != null && !accounts.isEmpty()) {
-            updateKinAccounts(accounts);
-        }
-    }
-
-    private void updateKinAccounts(List<KeyPair> storageAccounts) {
-        Map<String, KinAccountImpl> accountsMap = new HashMap<>();
-        for (KinAccountImpl kinAccountImpl : kinAccounts) {
-            accountsMap.put(kinAccountImpl.getPublicAddress(), kinAccountImpl);
-        }
-
-        List<KinAccountImpl> newKinAccountsList = new ArrayList<>();
-        for (KeyPair account : storageAccounts) {
-            KinAccountImpl inMemoryKinAccount = accountsMap.get(account.getAccountId());
-            if (inMemoryKinAccount != null) {
-                newKinAccountsList.add(inMemoryKinAccount);
-            } else {
-                newKinAccountsList.add(createNewKinAccount(account));
-            }
-        }
-        kinAccounts = newKinAccountsList;
-    }
-
-    private void validateAppId(String appId) {
-        if (appId == null || appId.equals("")) {
-            Log.w("KinClient","WARNING: KinClient instance was created without a proper application ID. Is this what you intended to do?");
-        }
-        else if (!appId.matches("[a-zA-Z0-9]{3,4}")) {
-            throw new IllegalArgumentException("appId must contain only upper and/or lower case letters and/or digits and that the total string length is between 3 to 4.\n" +
-                    "for example 1234 or 2ab3 or cd2 or fqa, etc.");
-        }
     }
 
     /**
@@ -147,8 +78,7 @@ public class KinClient {
      */
     public @NonNull
     KinAccount addAccount() throws CreateAccountException {
-        KeyPair account = keyStore.newAccount();
-        return addKeyPair(account);
+        return kinClientInternal.addAccount();
     }
 
     /**
@@ -161,29 +91,7 @@ public class KinClient {
     @NonNull
     public KinAccount importAccount(@NonNull String exportedJson, @NonNull String passphrase)
         throws CryptoException, CreateAccountException, CorruptedDataException {
-        KeyPair account = keyStore.importAccount(exportedJson, passphrase);
-        KinAccount kinAccount = getAccountByPublicAddress(account.getAccountId());
-        return kinAccount != null ? kinAccount : addKeyPair(account);
-    }
-
-    @Nullable
-    private KinAccount getAccountByPublicAddress(String accountId) { //TODO we should make this method public
-        loadAccounts();
-        KinAccount kinAccount = null;
-        for (int i = 0; i < kinAccounts.size(); i++) {
-            final KinAccount account = kinAccounts.get(i);
-            if (accountId.equals(account.getPublicAddress())) {
-                kinAccount = account;
-            }
-        }
-        return kinAccount;
-    }
-
-    @NonNull
-    private KinAccount addKeyPair(KeyPair account) {
-        KinAccountImpl newAccount = createNewKinAccount(account);
-        kinAccounts.add(newAccount);
-        return newAccount;
+       return kinClientInternal.importAccount(exportedJson, passphrase);
     }
 
     /**
@@ -192,26 +100,21 @@ public class KinClient {
      * @return the account at the input index or null if there is no such account
      */
     public KinAccount getAccount(int index) {
-        loadAccounts();
-        if (index >= 0 && kinAccounts.size() > index) {
-            return kinAccounts.get(index);
-        }
-        return null;
+        return kinClientInternal.getAccount(index);
     }
 
     /**
      * @return true if there is an existing account
      */
     public boolean hasAccount() {
-        return getAccountCount() != 0;
+        return kinClientInternal.hasAccount();
     }
 
     /**
      * Returns the number of existing accounts
      */
     public int getAccountCount() {
-        loadAccounts();
-        return kinAccounts.size();
+       return kinClientInternal.getAccountCount();
     }
 
     /**
@@ -220,36 +123,18 @@ public class KinClient {
      * @throws DeleteAccountException in case of a delete account exception while trying to delete the account
      */
     public boolean deleteAccount(int index) throws DeleteAccountException {
-        boolean deleteSuccess = false;
-        if (index >= 0 && getAccountCount() > index) {
-            String accountToDelete = kinAccounts.get(index).getPublicAddress();
-            keyStore.deleteAccount(accountToDelete);
-            KinAccountImpl removedAccount = kinAccounts.remove(index);
-            removedAccount.markAsDeleted();
-            deleteSuccess = true;
-        }
-        return deleteSuccess;
+        return kinClientInternal.deleteAccount(index);
     }
 
     /**
      * Deletes all accounts.
      */
     public void clearAllAccounts() {
-        keyStore.clearAllAccounts();
-        for (KinAccountImpl kinAccount : kinAccounts) {
-            kinAccount.markAsDeleted();
-        }
-        kinAccounts.clear();
+        kinClientInternal.clearAllAccounts();
     }
 
     public Environment getEnvironment() {
-        return environment;
-    }
-
-    @NonNull
-    private KinAccountImpl createNewKinAccount(KeyPair account) {
-        return new KinAccountImpl(account, backupRestore, transactionSender,
-                accountInfoRetriever, blockchainEventsCreator);
+        return kinClientInternal.getEnvironment();
     }
 
     /**
@@ -259,12 +144,7 @@ public class KinClient {
      * @return {@code Request<Integer>} - the minimum fee.
      */
     public Request<Long> getMinimumFee() {
-        return new Request<>(new Callable<Long>() {
-            @Override
-            public Long call() throws Exception {
-                return generalBlockchainInfoRetriever.getMinimumFeeSync();
-            }
-        });
+        return kinClientInternal.getMinimumFee();
     }
 
     /**
@@ -275,11 +155,11 @@ public class KinClient {
      * @return the minimum fee.
      */
     public long getMinimumFeeSync() throws OperationFailedException {
-        return generalBlockchainInfoRetriever.getMinimumFeeSync();
+        return kinClientInternal.getMinimumFeeSync();
     }
 
     public String getAppId() {
-        return appId;
+        return kinClientInternal.getAppId();
     }
 
     public String getStoreKey() {
