+++ kin-sdk/kin-sdk-lib/src/test/java/kin/sdk/WhitelistServiceForTest.java
